refactor(modules): consolidate the guardrails subsystem into src/modules/guardrails/
Relocate the guardrails subsystem — guardrails.c, guardrails_orchestrator.c,
guardrails_action_audit.c, guardrails_tdd.c, guardrails_semantic.c,
guardrails_blast_radius.c, and cmd_guardrails.c — into
src/modules/guardrails/, consolidating their headers (from src/headers/ and
the src root) alongside them, matching the economizer/memory/workspace module
layout.

Add -Imodules/guardrails to the compiler include path and re-prefix source
and object references across src/Makefile (DATA_SRCS, CMD_SRCS, the
SERVER_DATA filter-out and object override, BENCH_OBJS), src/tests/Rules.mk,
and CMakeLists.txt (sources plus the six target include-dir lists). Fix the
one path-qualified test include (test_guardrails_semantic.c), update the
guardrails.h path in scripts/check_tier_deps.sh, and refresh the MANUAL and
OWNERS doc references.
