+++ src/modules/guardrails/guardrails_action_audit.c
