+++ src/modules/guardrails/guardrails_internal.h
